Устранил некоторые недостатки
