+++ package.json
@@ -20,6 +20,6 @@
     "node": "12"
   },
   "dependencies": {
-    "dayjs": "^1.10.6"
+    "dayjs": "1.10.6"
   }
 }
